move logic to vue-components
