+++ packages/vue-components/src/Navbar.vue
@@ -77,10 +77,41 @@
       toggleCollapse (e) {
         e && e.preventDefault()
         this.collapsed = !this.collapsed
+      },
+      highlightCurrentPage() {
+        function hasMatchingUrl(elements, currPage) {
+          if (!elements || !elements.children) {
+            return false;
+          }
+          // Only check <a> leaf nodes
+          if (elements.children.length === 0) {
+            if (elements.href) {
+              if (elements.href === currPage) {
+                return true;
+              }
+            }
+            return false;
+          }
+          // otherwise, check all children recursively
+          return Array.from(elements.children).some(node => hasMatchingUrl(node, currPage));
+        }
+        document.addEventListener('DOMContentLoaded', () => {
+          const navLinks = document.querySelectorAll('.navbar .navbar-nav .nav-link');
+          if (!navLinks) {
+            return;
+          }
+          const currPage = window.location.href;
+          Array.from(navLinks).forEach((node) => {
+            if (hasMatchingUrl(node, currPage)) {
+              node.classList.add('current');
+            }
+          });
+        });
       }
     },
     created () {
       this._navbar = true
+      this.highlightCurrentPage();
     },
     mounted () {
       let $dropdown = $('.dropdown>[data-toggle="dropdown"]',this.$el).parent()
